fresh rg/requirements.txt
discord.py>=2.3.2
motor>=3.3.1
python-dotenv>=1.0.0
fastapi>=0.95.0
uvicorn>=0.21.1
python-multipart>=0.0.6
aiohttp>=3.8.4
jinja2>=3.1.2
